Condensed row: split into file and cleanups (#5743)
* split and cleanup

* clean up
